basic structure
Added the js Library and Book objects and a method for adding Books to the Library's inventory

Laid out the general structure for the html/css part. Form on top with relevant input fields, display on the bottom that shows cover art for a selected book and a list of books.

The goal is that the list of books at the bottom will update with clickable anchors that when clicked will display all the book's information and cover art.